Move tests to a more suitable class
